Reworked the XML parsing regexps.
The XML parsing code was made a little bit more powerful so now we can parse
node attributes and other information. Messages were switched to use this new
parser, and they also now return the text contained within the <message> nodes
(as opposed to just what was in the attributes).